fix: bundle sandbox libraries as static imports for compiled binary
Extract worker runtime into startWorkerRuntime() that accepts preloaded
modules. The server's custom worker entry statically imports @libpdf/core
so it gets bundled into the Bun binary as a second entrypoint.

- Add worker-runtime.ts with preloadedModules fallback pattern
- Add sandbox-worker.ts in server as the compiled binary's worker entry
- Make createWorkerSandbox accept a configurable workerUrl
- Remove dist/node_modules from electron-builder extraResources
