+++ apps/desktop/electron-builder.config.ts
@@ -43,11 +43,6 @@ const config: Configuration = {
       filter: ['stitch-server*'],
     },
     {
-      from: '../../packages/server/dist/node_modules',
-      to: 'node_modules',
-      filter: ['**/*'],
-    },
-    {
       from: '../../packages/server/drizzle',
       to: 'drizzle',
       filter: ['**/*'],
