+++ packages/sandbox/src/index.ts
@@ -1,2 +1,3 @@
 export { createWorkerSandbox } from './driver.js';
+export { startWorkerRuntime } from './worker-runtime.js';
 export type { IsolateDriver, IsolateOptions, ToolBinding } from './types.js';
