+++ packages/server/package.json
@@ -7,8 +7,8 @@
   "scripts": {
     "test": "bun test",
     "dev": "NODE_ENV=development bun --watch src/index.ts",
-    "build": "bun build src/index.ts --compile --minify --outfile dist/stitch-server",
-    "build:x64": "bun build src/index.ts --compile --minify --target=bun-darwin-x64 --outfile dist/stitch-server",
+    "build": "bun build src/index.ts src/code-mode/sandbox-worker.ts --compile --minify --outfile dist/stitch-server",
+    "build:x64": "bun build src/index.ts src/code-mode/sandbox-worker.ts --compile --minify --target=bun-darwin-x64 --outfile dist/stitch-server",
     "typecheck": "tsc --noEmit",
     "lance:migration:new": "bun scripts/new-lance-migration.ts"
   },
